Tackle PR comment
